[feat] handle descriptor update
